error fix

+++ Lab-10/Sketchpad/src/main/java/Shape.java
@@ -1,5 +1,5 @@
 public abstract class Shape {
-    public abstract int area();
+    public abstract double area();
     public abstract String describe();
 
     public abstract void draw(Canvas canvas);
